docker/Makefile: Make sure custom JSGLR_VERSION only needs to be built once
By pulling jsglr with the right version before calling ./b build java,
we don't need to manually call the maven scripts.
The build-jsglr2 job still exists, just in case you still want to use it,
but it's no longer included in the 'all' target.
Only the benchmark and measure projects need to be built manually,
because they are not included in ./b build java.